fix incorrect masking
storing 36 bits of data in a 32 bit integer doesn't work so well

+++ src/Ryujinx.Cpu/AddressTable.cs
@@ -238,7 +238,7 @@ namespace ARMeilleure.Common
             {
                 TEntry* page = GetPage(address);
 
-                int index = Levels[^1].GetValue(address);
+                long index = Levels[^1].GetValue(address);
 
                 EnsureMapped((IntPtr)(page + index));
